backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
vtracer==0.6.15
rembg==2.0.73
Pillow>=12.1.0
slowapi==0.1.9
